Resolvendo problemas citados no rubocop
