+++ app/controllers/api/v1/base_user_auth_controller.rb
@@ -2,7 +2,8 @@
 
 module Api
   module V1
-    class ApplicationController < ActionController::API
+    # Base methods context API
+    class BaseUserAuthController < ActionController::API
       rescue_from CanCan::AccessDenied do |_exception|
         render json: { message: 'Access denied' }, status: 403
       end
@@ -12,27 +13,25 @@ module Api
       end
 
       def auth_header
-        # { Authorization: 'Bearer <token>' }
         request.headers['Authorization']
       end
 
       def decoded_token
-        if auth_header
-          token = auth_header.split[1]
-          # header: { 'Authorization': 'Bearer <token>' }
-          begin
-            JWT.decode(token, 's3cr3t', true, algorithm: 'HS256')
-          rescue JWT::DecodeError
-            nil
-          end
+        return unless auth_header
+
+        token = auth_header.split[1]
+        begin
+          JWT.decode(token, 's3cr3t', true, algorithm: 'HS256')
+        rescue JWT::DecodeError
+          nil
         end
       end
 
       def logged_in_user
-        if decoded_token
-          user_id = decoded_token[0]['user_id']
-          @user = User.find_by(id: user_id)
-        end
+        return unless decoded_token
+
+        user_id = decoded_token[0]['user_id']
+        @user = User.find_by(id: user_id)
       end
 
       def logged_in?
