--- app/controllers/api/v1/registrations_controller.rb
@@ -1,26 +1,0 @@
-# frozen_string_literal: true
-
-module Api
-  module V1
-    # class RegistrationsController < Devise::RegistrationsController
-    #   # protect_from_forgery unless: -> { request.format.json? }
-    #   respond_to :json
-
-    #   private
-
-    #   def respond_with(resource, _opts = {})
-    #     register_success && return if resource.persisted?
-
-    #     register_failed
-    #   end
-
-    #   def register_success
-    #     render json: { message: 'Signed up sucessfully.' }
-    #   end
-
-    #   def register_failed
-    #     render json: { message: 'Something went wrong.' }
-    #   end
-    # end
-  end
-end
