--- app/controllers/api/v1/sessions_controller.rb
@@ -1,30 +1,0 @@
-# frozen_string_literal: true
-
-module Api
-  module V1
-    # class SessionsController < Devise::SessionsController
-    #   skip_before_action :verify_authenticity_token
-    #   respond_to :json
-
-    #   private
-
-    #   def respond_with(_resource, _opts = {})
-    #     render json: { message: 'You are logged in. ' }, status: :ok
-    #   end
-
-    #   def respond_to_on_destroy
-    #     log_out_success && return if current_user
-
-    #     log_out_failure
-    #   end
-
-    #   def log_out_success
-    #     render json: { message: 'You are logged out.' }, status: :ok
-    #   end
-
-    #   def log_out_failure
-    #     render json: { message: 'Hmm nothing happened.' }, status: :unauthorized
-    #   end
-    # end
-  end
-end
