+++ app/controllers/application_controller.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+# Application methods context
 class ApplicationController < ActionController::Base
   rescue_from CanCan::AccessDenied do
     flash[:error] = 'Access denied!'
