+++ app/controllers/roles_controller.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+# Roles actions
 class RolesController < ApplicationController
   before_action :set_role, only: %i[show edit update destroy]
