+++ app/controllers/users_controller.rb
@@ -1,29 +1,25 @@
 # frozen_string_literal: true
 
+# User actions
 class UsersController < ApplicationController
   before_action :set_user, only: %i[show edit update destroy]
   before_action :authenticate_user!
   load_and_authorize_resource
 
-  # GET /users or /users.json
   def index
     @users = User.all
   end
 
-  # GET /users/1 or /users/1.json
   def show; end
 
-  # GET /users/new
   def new
     @user = User.new
   end
 
-  # GET /users/1/edit
   def edit; end
 
-  # POST /users or /users.json
   def create
-    @user = User.new(user_params)
+    @user = CreateUserService.new(user_params).call
 
     respond_to do |format|
       if @user.save
@@ -36,7 +32,6 @@ class UsersController < ApplicationController
     end
   end
 
-  # PATCH/PUT /users/1 or /users/1.json
   def update
     respond_to do |format|
       if @user.update(user_params)
@@ -49,7 +44,6 @@ class UsersController < ApplicationController
     end
   end
 
-  # DELETE /users/1 or /users/1.json
   def destroy
     @user.destroy
 
@@ -61,14 +55,12 @@ class UsersController < ApplicationController
 
   private
 
-  # Use callbacks to share common setup or constraints between actions.
   def set_user
     return if params[:id].to_i <= 0
 
     @user = User.find(params[:id])
   end
 
-  # Only allow a list of trusted parameters through.
   def user_params
     params.require(:user).permit(
       :email,
