--- app/helpers/api/v1/users_helper.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Api
-  module V1
-    module UsersHelper
-    end
-  end
-end
