+++ app/models/application_record.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+# ApplicationRecord
 class ApplicationRecord < ActiveRecord::Base
   primary_abstract_class
 end
