+++ app/models/role.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
+# role model
 class Role < ApplicationRecord
 end
