+++ app/views/users/_form.html.erb
@@ -17,6 +17,11 @@
   </div>
 
   <div>
+    <%= form.label :email, style: "display: block" %>
+    <%= form.text_field :email %>
+  </div>
+
+  <div>
     <%= form.label :role_id, style: "display: block" %>
     <%= form.text_field :role_id %>
   </div>
