--- spec/helpers/roles_helper_spec.rb
@@ -1,17 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-
-# Specs in this file have access to a helper object that includes
-# the RolesHelper. For example:
-#
-# describe RolesHelper do
-#   describe "string concat" do
-#     it "concats two strings with spaces" do
-#       expect(helper.concat_strings("this","that")).to eq("this that")
-#     end
-#   end
-# end
-RSpec.describe RolesHelper, type: :helper do
-  pending "add some examples to (or delete) #{__FILE__}"
-end
