+++ spec/services/create_user_service_spec.rb
@@ -1,0 +1,31 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe CreateUserService, type: :service do
+  include ActiveJob::TestHelper
+
+  context 'when has valid atributes to user' do
+    let(:role) { create(:role) }
+    let(:user_params) do
+      {
+        name: 'Flavio Avila',
+        email: 'flavio.avila.silva@outlook.com',
+        password: '123456',
+        password_confirmation: '123456',
+        role_id: role.id
+      }
+    end
+
+    subject { CreateUserService.new(user_params).call }
+
+    it 'create the user and send the welcome e-mail' do
+      clear_enqueued_jobs
+      subject
+
+      user = User.find_by(email: user_params[:email])
+      expect(user).not_to be_nil
+      expect(enqueued_jobs.size).to eq(1)
+    end
+  end
+end
