+++ spec/views/users/index.html.erb_spec.rb
@@ -9,7 +9,8 @@ RSpec.describe 'users/index', type: :view do
     user_admin = User.create!({ name: 'Admin', email: 'admin@admin.com', password: 'admin1',
                                 password_confirmation: 'admin1', role_id: role_admin.id })
     another_user = User.create!({ name: 'Another_user', email: 'another_user@another_user.com',
-                                  password: 'another_user', password_confirmation: 'another_user', role_id: role_admin.id })
+                                  password: 'another_user', password_confirmation: 'another_user',
+                                  role_id: role_admin.id })
 
     assign(:users, [user_admin, another_user])
   end
